Now work in double

+++ src/main/java/com/example/calculatorwithinterface/Main.java
@@ -12,36 +12,18 @@ import java.util.Scanner;
 public class Main extends Application {
 
     public static void main(String[] args) {
-//        Scanner in = new Scanner(System.in);
-//        String str = in.next();
-//
-//        String[] s = str.split("[+\\-/*]");
-//        String[] symb = Arrays.stream(str.split("\\d")).filter(x -> !x.isEmpty()).toArray(String[]::new);
-//        int[] newStr = new int[] {Integer.parseInt(s[0]), Integer.parseInt(s[1])};
-//        int res = newStr[0];
-//        for (int i = 1; i < s.length; i++) {
-//            newStr[0] = res;
-//            newStr[1] = Integer.parseInt(s[i]);
-//            switch (symb[i - 1]) {
-//                case "+" -> res = newStr[0] + newStr[1];
-//                case "-" -> res = newStr[0] - newStr[1];
-//                case "/" -> res = newStr[0] / newStr[1];
-//                case "*" -> res = newStr[0] * newStr[1];
-//            }
-//        }
-//        System.out.println(res);
         launch(args);
     }
 
-    public float Calculation(String str) {
+    public double Calculation(String str) {
         String[] s = str.split("[+\\-/*]");
         if (s.length < 2) {
-            return Float.parseFloat(s[0]);
+            return Double.parseDouble(s[0]);
 
         } else {
             String[] symb = Arrays.stream(str.split("[.\\d]")).filter(x -> !x.isEmpty()).toArray(String[]::new);
-            float[] newStr = new float[]{Float.parseFloat(s[0]), Float.parseFloat(s[1])};
-            float res = newStr[0];
+            double[] newStr = new double[]{Float.parseFloat(s[0]), Float.parseFloat(s[1])};
+            double res = newStr[0];
             for (int i = 1; i < s.length; i++) {
                 newStr[0] = res;
                 newStr[1] = Float.parseFloat(s[i]);
